+++ fj-doc-playground-quarkus/src/test/java/org/fugerit/java/doc/playground/GenerateRestTest.java
@@ -37,9 +37,10 @@ class GenerateRestTest {
 	
     @Test
     void testGenerateDocument() {   	
-    	int[] testId = { 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+    	int[] testId = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
     	for ( int k=0; k<testId.length; k++ ) {
-    		this.testWorker( "/generate/document", "generate/test_generate_input_0"+testId[k]+".json" );
+    		String current = String.valueOf( testId[k]<10 ? "0"+testId[k] : testId[k] ); 
+    		this.testWorker( "/generate/document", "generate/test_generate_input_"+current+".json" );
     	}   
     }