+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [playground-quarkus] xsl-fo output (as of mod-fop format)
+
 ### Changed
 
 - [playground-quarkus] quarkus-version set to 3.6.1